custom_sable/requirements.txt
# Core dependencies for custom Sable implementation
torch>=1.9.0
numpy>=1.21.0
einops>=0.4.0
pyyaml>=6.0
tqdm>=4.62.0
tensorboard>=2.7.0
typing-extensions>=4.0.0